+++ integration/samples/custom-node-resolve/src/baz/baz.component.ts
@@ -1,0 +1,8 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'custom-baz',
+  templateUrl: './baz.component.html',
+  styleUrls: ['./baz.component.less', './baz.component.css']
+})
+export class BazComponent {}