+++ integration/samples/custom-node-resolve/src/public_api.ts
@@ -1,0 +1,6 @@
+export * from './foo/foo.component';
+export * from './bar/bar.component';
+export * from './baz/baz.component';
+export * from './foo-bar/foo-bar.component';
+export * from './less-baz/less-baz.component';
+export * from './custom.module';