+++ src/lib/flatten/rollup.ts
@@ -6,6 +6,7 @@ import * as log from '../util/log';
 import { BuildStep } from '../deprecations';
 import { externalModuleIdStrategy } from './external-module-id-strategy';
 import { umdModuleIdStrategy } from './umd-module-id-strategy';
+import { inspect } from 'util';
 
 export type BundleFormat = __rollup.Format;
 
@@ -16,18 +17,23 @@ export interface RollupOptions {
   dest: string;
   umdModuleIds?: { [key: string]: string };
   embedded?: string[];
+  externalOpts?: any;
 }
 
 /** Runs rollup over the given entry file, writes a bundle file. */
 export async function rollup(opts: RollupOptions): Promise<void> {
   log.debug(`rollup (v${__rollup.VERSION}) ${opts.entry} to ${opts.dest} (${opts.format})`);
+  log.debug(`rollup custom options ${inspect(opts.externalOpts)}`);
+
+  const nodeResolveOpts =
+    opts.externalOpts && opts.externalOpts.nodeResolve ? opts.externalOpts.nodeResolve : { jsnext: true, module: true };
 
   // Create the bundle
   const bundle: __rollup.Bundle = await __rollup.rollup({
     context: 'this',
     external: moduleId => externalModuleIdStrategy(moduleId, opts.embedded || []),
     input: opts.entry,
-    plugins: [nodeResolve({ jsnext: true, module: true }), commonJs()],
+    plugins: [nodeResolve(nodeResolveOpts), commonJs()],
     onwarn: warning => {
       if (warning.code === 'THIS_IS_UNDEFINED') {
         return;